fix(fourier_2d_rnn.py): correct grid structure for RNN
The previous version creates data such that the space coordinates are at the starting of the feature list (12 features- 2 space and 10 field). Therefore when we do: 

xx = torch.cat((xx[..., step:-2], im,
                            gridx.repeat([batch_size, 1, 1, 1]), gridy.repeat([batch_size, 1, 1, 1])), dim=-1)

it trims the field features instead of the spatial features, but this is not what we want. We want to trim the spatial features, hence pushing them back helps.